+++ merged-packages/eth-json-rpc-middleware/src/index.ts
@@ -2,7 +2,7 @@ export * from './block-cache';
 export * from './block-ref-rewrite';
 export * from './block-ref';
 export * from './block-tracker-inspector';
-export * from './fetch';
+export { createFetchMiddleware } from './fetch';
 export * from './inflight-cache';
 export type {
   GetCallsStatusHook,