+++ merged-packages/eth-json-rpc-middleware/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- **BREAKING:** `createFetchMiddleware` no longer takes `fetch`, `btoa`, `rpcUrl`, and `originHttpHeaderKey` ([#402](https://github.com/MetaMask/eth-json-rpc-middleware/pull/402))
+  - The existing signature (`rpcService` and `options`) is now the only way to use this function; please use that instead.
+
+### Removed
+
+- **BREAKING:** Remove `PayloadWithOrigin` type ([#402](https://github.com/MetaMask/eth-json-rpc-middleware/pull/402))
+  - There is no equivalent.
+- **BREAKING:** Remove `createFetchConfigFromReq` function ([#402](https://github.com/MetaMask/eth-json-rpc-middleware/pull/402))
+  - There is no equivalent.
+
 ## [17.1.0]
 
 ### Added